refactor: 토큰 발급 기능 모듈 이동
