+++ in-adapter-api/src/main/java/gohigher/auth/response/TokenResponse.java
@@ -1,0 +1,12 @@
+package gohigher.auth.response;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public class TokenResponse {
+
+	private final String accessToken;
+	private final String role;
+}
