+++ in-adapter-api/src/main/java/gohigher/auth/support/AuthorizationExtractor.java
@@ -1,4 +1,4 @@
-package gohigher.auth;
+package gohigher.auth.support;
 
 import java.util.Enumeration;
