+++ in-adapter-api/src/main/java/gohigher/auth/support/CookieProvider.java
@@ -1,4 +1,4 @@
-package gohigher.jwt.support;
+package gohigher.auth.support;
 
 import java.time.Duration;
