+++ security/src/main/java/gohigher/jwt/RoleGrantedAuthority.java
@@ -1,4 +1,4 @@
-package gohigher.jwt.support;
+package gohigher.jwt;
 
 import java.io.Serial;
