+++ security/src/main/resources/application-auth.yml
@@ -24,7 +24,7 @@ spring:
 
 oauth2:
   success:
-    redirect_uri: 'http://localhost:3000'
+    redirect_uri: 'http://localhost:8080'
 
 token:
   request:
